Moved Command related functions to "src/commands.c"
