+++ build.bash
@@ -1,3 +1,3 @@
 #! /bin/bash
-gcc src/main.c src/builtins.c src/parsing.c src/signals.c src/execute.c -o clobber.out -g3
+gcc src/main.c src/builtins.c src/parsing.c src/signals.c src/execute.c src/commands.c -o clobber.out -g3
 ./clobber.out
